# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "GPL 3",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        }
    },
    "paths": {
        "/api/v1/user/edit_story/{story_id}": {
            "put": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Edit an existing story with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "story",
                    "user"
                ],
                "summary": "Edit a story",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Story ID to be edited",
                        "name": "story_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Story details to be edited",
                        "name": "story",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTO.StoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DTO.StoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user/delete": {
            "delete": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Delete the user account associated with the provided JWT token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Delete user account",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DTO.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/delete_story/{id}": {
            "delete": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Delete a story with the provided ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "story",
                    "user"
                ],
                "summary": "Delete a story",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Story ID to be deleted",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DTO.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    }
                },
                "x-order": 8
            }
        },
        "/user/login": {
            "post": {
                "description": "Authenticate a user with their credentials and retrieve an access token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "User login credentials",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTO.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/DTO.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/new_story": {
            "post": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Create a new story with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user",
                    "story"
                ],
                "summary": "Create a new story",
                "parameters": [
                    {
                        "description": "New story details",
                        "name": "story",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTO.CreateStoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/DTO.StoryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "Register a new user with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration details",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTO.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DTO.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/stories": {
            "get": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "description": "Retrieve all posted stories of the user based on specified options.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "story",
                    "user"
                ],
                "summary": "Get all posted stories of the user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sort the stories by a field (e.g., 'created')",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit the number of returned stories",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset the returned stories",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter stories by start date (2006-01-02T15:04:05Z07:00)",
                        "name": "from_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter stories by end date (2006-01-02T15:04:05Z07:00)",
                        "name": "to_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DTO.StoriesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/DTO.ErrorResponse"
                        }
                    }
                },
                "x-order": 5
            }
        }
    },
    "definitions": {
        "DTO.CreateStoryRequest": {
            "type": "object",
            "properties": {
                "attachedFile": {
                    "type": "string",
                    "example": "http://example.com/myfile.pdf"
                },
                "backgroundColor": {
                    "type": "string",
                    "example": "#RRGGBB"
                },
                "backgroundImage": {
                    "type": "string",
                    "example": "http://example.com/myimage.jpg"
                },
                "externalWebLink": {
                    "type": "string",
                    "example": "http://example.com"
                },
                "fromTime": {
                    "type": "string",
                    "example": "2020-01-01T00:00:00Z"
                },
                "isShareable": {
                    "type": "boolean",
                    "example": true
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/DTO.Status"
                        }
                    ],
                    "example": "private/public"
                },
                "storyName": {
                    "type": "string",
                    "example": "My first story"
                },
                "toTime": {
                    "type": "string",
                    "example": "2020-12-31T23:59:59Z"
                }
            }
        },
        "DTO.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "DTO.LoginResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                }
            }
        },
...
        "DTO.RegisterRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "DTO.Status": {
            "type": "string",
            "enum": [
                "private",
                "public"
            ],
            "x-enum-varnames": [
                "StatusPrivate",
                "StatusPublic"
            ]
        },
        "DTO.StoriesResponse": {
            "type": "object",
            "properties": {
                "stories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DTO.StoryResponse"
                    }
                }
            }
        },
        "DTO.StoryRequest": {
            "type": "object",
            "properties": {
                "attachedFile": {
                    "type": "string"
                },
                "backgroundColor": {
                    "type": "string"
                },
                "backgroundImage": {
                    "type": "string"
                },
                "creatorUserId": {
                    "type": "integer"
                },
                "externalWebLink": {
                    "type": "string"
                },
                "fromTime": {
                    "type": "string"
                },
                "isShareable": {
                    "type": "boolean"
                },
                "status": {
                    "$ref": "#/definitions/DTO.Status"
                },
                "storyID": {
                    "type": "integer"
                },
                "storyName": {
                    "type": "string"
                },
                "toTime": {
                    "type": "string"
                }
            }
        },
        "DTO.StoryResponse": {
            "type": "object",
            "properties": {
                "attachedFile": {
                    "type": "string"
                },
                "backgroundColor": {
                    "type": "string"
                },
                "backgroundImage": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "creatorUserId": {
                    "type": "integer"
                },
                "externalWebLink": {
                    "type": "string"
                },
                "fromTime": {
                    "type": "string"
                },
                "isShareable": {
                    "type": "boolean"
                },
                "relevance": {
                    "type": "boolean"
                },
                "status": {
                    "$ref": "#/definitions/DTO.Status"
                },
                "storyID": {
                    "type": "integer"
                },
                "storyName": {
                    "type": "string"
                },
                "toTime": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "DTO.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "GuestAuth": {
            "description": "Just put the guest api key in header.",
            "type": "apiKey",
            "name": "X-Guest-Token",
            "in": "header"
        },
        "JwtAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}